[#17] fix: search not working, move data file to src
